Add GameCommand as user input into game world.
A request from a user is turned into a command, including generating the
uuid that is the clients "reference" to the unit. This lets the client
know have a reference before it is added into the world.
The websocket communicates with with the main thread through a channel.
This channel does not block the sender, and the reciver can empty
the channels buffer before each tick. Theese commands are then inserted
into the game world and handled by the handle_commands system.